+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
-## [2.6.0] - 2024-04-18
+## [2.5.2] - 2024-04-18
 - Updated mongo-lib to v4.3.0
 - Added TLS capability
 - Set pymongo to 3.12.3 for Python 2 and Python 3.
+
+### Fixed
+- run_compact:  Moved the change to database to before getting a table list.
 
 ### Changed
 - Set pymongo to 3.12.3 for Python 2 and Python 3.